+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ### v0.3.0
-- :bulb: Added `NumEmailSent` score criteria
+- :bulb: Added `NumEmailsSent` score criteria
 - :hammer: Fixed `ZazoActivity` score criteria
 - :bulb: Added update info button to `contact#show` page
 - :bulb: Added links to other services to `contact#show` page